% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_numerical_fit_TIME.R
\name{numerical_fit_TIME}
\alias{numerical_fit_TIME}
\title{Numerically fit the Breakthrough Curves - Time Methodology}
\usage{
numerical_fit_TIME(
  time,
  CCo,
  flask_volume,
  column_length,
  column_diameter,
  soil_density,
  particle_density
)
}
\arguments{
\item{time}{- Time elapsed between each experimental observation - LIST.}

\item{CCo}{- Relative concentration between each experimental observation - LIST.}

\item{flask_volume}{- Volume of the solute collection flask (cm³) - Numerical Input.}

\item{column_length}{- total length of soil column (cm) - Numerical Input.}

\item{column_diameter}{- total diameter of soil column (cm) - Numerical Input.}

\item{soil_density}{- Soil density for the soil under analysis (g. cm-3) - Numerical Input.}

\item{particle_density}{- Particle density for the soil under analysis (g. cm-3) - Numerical Input.}
}
\value{
Returns a table with the experimental information, numerical fit of the relative concentration, transport parameters, and RMSE.
}
\description{
Performs the numerical adjustment of the relative concentration of solutes through the time-based methodology - check inputs; and returns a dataframe with the fitted values and transport parameters.
}
\examples{
MRE <- xlsx::read.xlsx("./inst/app/www/MRE.xlsx",sheetIndex = 1) \%>\% tidyr::as_tibble()
fitted_MRE <- numerical_fit_TIME(MRE$Time, MRE$CCo, 23, 20, 5.2, 1.46, 2.57)

}
