% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_BTC.R
\name{plot_BTC}
\alias{plot_BTC}
\title{Plot the fitted data}
\usage{
plot_BTC(PV, CCo, Fitted)
}
\arguments{
\item{PV}{Pore Volume measured between each experimental observation (as exported in one of the above functions).}

\item{CCo}{Relative concentration between each experimental observation (as exported in one of the above functions).}

\item{Fitted}{Numerical fit of the relative concentration (as exported in one of the above functions).}
}
\value{
Returns the plot of the BTCs adjusted by BTCfit
}
\description{
Plot the performed Fit of the experimental data.
}
\examples{
MRE <- xlsx::read.xlsx("./inst/app/www/MRE.xlsx",sheetIndex = 1) \%>\% tidyr::as_tibble()
fitted_MRE <- numerical_fit_PV(MRE$Pore_Volume, MRE$CCo, 0.42, 20, 5.2)
BTC <- plot_BTC(fitted_MRE$PV,MRE$CCo,fitted_MRE$Fitted)

}
